hackwithus/requirements.txt
urllib3==1.23

Django==2.1.3   

google-api-core==1.5.2    

google-auth==1.6.1    

google-cloud-automl==0.1.1    

googleapis-common-protos==1.5.5    

Pillow==5.2.0    

PyHamcrest==1.9.0

requests==2.19.1
grpcio==1.16.0